langchain-langraph-workspace/langchain_project/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.3.67
langchain-openai>=0.1.0
langchain-text-splitters>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pypdf>=3.0.0
tenacity>=8.1.0
python-multipart>=0.0.9
